VP1-226-Merge
Merge, to match VLAT_Demo changes
