+++ Assets/VERA/VLAT_LocomotionAccessibilityToolkit/Prefabs.meta
@@ -1,5 +1,6 @@
 fileFormatVersion: 2
-guid: 38710a707223e2244a2f589f20731968
+guid: 2fa3cd9410f02f24ebdf37856f8e248f
+folderAsset: yes
 DefaultImporter:
   externalObjects: {}
   userData:
