+++ Assets/VERA/VLAT_LocomotionAccessibilityToolkit/Scripts/VLAT_DoorInteractable.cs
@@ -85,73 +85,6 @@ public class VLAT_DoorInteractable : MonoBehaviour
     {
         StopCoroutine(RotateToTarget());
         StartCoroutine(RotateToTarget());
-        /*
-        // Stop velocity
-        rb.velocity = Vector3.zero;
-        rb.angularVelocity = Vector3.zero;
-
-        // If the door is two-way, it has 3 states (pull open, closed, push open)
-        // Swap between states, going 0, 1, 2, 3, 0, 1, 2, 3...)
-        // (Which represents closed from open pulled, open pushed, closed from open pushed, open pulled, closed from open pulled...)
-        if (twoWayDoor)
-        {
-            if (doorType == DoorType.Hinged)
-            {
-                float targetAngle = 0f;
-                switch (doorState)
-                {
-                    case 0:
-                        targetAngle = pushedOpenHingeAngle - doorHingeJoint.angle;
-                        break;
-                    case 1:
-                        targetAngle = closedHingeAngle - doorHingeJoint.angle;
-                        break;
-                    case 2:
-                        targetAngle = pulledOpenHingeAngle - doorHingeJoint.angle;
-                        break;
-                    case 3:
-                        targetAngle = closedHingeAngle - doorHingeJoint.angle;
-                        break;
-                }
-
-                transform.LeanRotateAround(doorHingeJoint.axis, targetAngle, .5f).setEaseOutExpo();
-
-                doorState++;
-                if (doorState > 3)
-                    doorState = 0;
-            }
-            else
-            {
-                // TODO
-            }
-        }
-        // If door is not two way, swap between two states, 0 and 1 (closed and opened)
-        else
-        {
-            if (doorType == DoorType.Hinged)
-            {
-                float targetAngle = 0f;
-                switch (doorState)
-                {
-                    case 0:
-                        targetAngle = openHingeAngle - doorHingeJoint.angle;
-                        break;
-                    case 1:
-                        targetAngle = closedHingeAngle - doorHingeJoint.angle;
-                        break;
-                }
-
-                transform.LeanRotateAround(doorHingeJoint.axis, targetAngle, .5f).setEaseOutExpo();
-
-                doorState++;
-                if (doorState > 1)
-                    doorState = 0;
-            }
-            else
-            {
-                // TODO
-            }
-        }*/
 
     } // END OpenAndClose
