+++ Assets/VERA/VLAT_LocomotionAccessibilityToolkit/Scripts/VLAT_GrabInteractable.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 196ca9ce809ef1a40a04c44c6ac9600a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
